Added disk interface to qml
We can now create basic disk objects from a qml file/javascript
The disk creation interface is added to the yaml.cpp/h files
We can now add disks/partitions and link them all up to generate a yaml
file.
resizable partitions and encrypted partitions can't yet be added but
basic primary partitions are possible.
We can specify if the partitiontable should be build and if the table is
EFi or MBR